Moved Coordinator class out of WebViewSnapshotRefresher
